+++ packages/ui/src/utils/type.ts
@@ -2,10 +2,10 @@ export const isListen = (str: string) => {
   return /^on.*\$/.test(str);
 };
 
-export const isValue = (value: any) => {
+export const isValue = (value: object): value is { untrackedValue: unknown; value: unknown } => {
   try {
     return 'untrackedValue' in value && 'value' in value;
-  } catch (error) {
+  } catch {
     return false;
   }
 };